pyproject.toml
[tool.poetry]
name = "bk7231tools"
version = "1.4.0"
description = "Tools to interact with and analyze artifacts for BK7231 MCUs"
authors = ["Khaled Nassar <kmhnassar@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.5"
pycryptodome = {version = "^3.16.0", optional = true}
py-datastruct = "^1.0.0"

[tool.poetry.extras]
cli = ["pycryptodome"]

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
autoflake = "^1.4"

[tool.poetry.scripts]
bk7231tools = "bk7231tools:cli"
bktools = "bk7231tools:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
